换用新的layui界面风格

+++ receive-platform/src/main/java/cn/com/cyber/service/impl/AppInfoServiceImpl.java
@@ -86,6 +86,9 @@ public class AppInfoServiceImpl implements AppInfoService {
             appModel.setAppId(appId);
             appModelList.add(appModel);
         }
+        if(appModelList.isEmpty()){
+           return;
+        }
         appInfoMapper.deleteAppServiceByAppId(appId);
         appInfoMapper.saveAppServiceMore(appModelList);
     }
